Basis input follows general parameter conventions
Basis set information (if needed) must be entered in the 'basis'
dictionary (of class Parameterdata), following the syntax of the
input_parameters dictionary.  This is a pragmatic solution to the
complexity of basis-set specification in Siesta: blocks can be
entered in 'raw' form, allowing in principle the matching of any
standard input file. This raw form complicates the provenance tagging,
but it is a working solution for now.