Made kpoints structure optional on input
The plugin now assumes that a calculation without a 'kpoints' element is
a gamma-point run. A check is made with the 'gamma_only' key in the
'settings' dictionary. If the latter is 'False' an error is
raised. Note that this forces the presence of a 'gamma_only': True
key-value pair in the settings dictionary.

The test_siesta.py test script has been updated. Elements have been
moved around, and the attachment of the different elements to the
calculation object is done next to their definition, if applicable.