Fix a typo and kpoint logic in input plugin. Update second test
With the new logic, gamma_only must be True in the settings dictionary.
This exposed a bug in the construction of the k-points block.

The second test (test_siesta_sih.py) needs to specify the settings
dictionary.

To Do: Perhaps set a default of None for gamma_only, instead of False.
This would open up more possibilities.