Implement basic restart functionality
Removed QE-specific items and updated the interface.
For now, restarts just involve copying the DM file from
the parent calculation folder, and setting 'dm-use-save-dm' to True.
(Restarts of geometry-optimization calculations will be implemented
when trajectories are available)

Note that restarts of calculations in which several
species share the same pseudopotential are not yet supported, as the
'kinds' argument to 'use_pseudo' is not directly available from the
parent calculation (it could be taken from the structure).